refactor(docsite): extract shared groupComponents, reuse in sidenav and package grid
- New lib/groupComponents.ts with the canonical grouping logic
- DocsShell.buildComponentSidebar now delegates to groupComponents
- Package page uses same grouping, so cards match the sidenav exactly
- Removes duplicated grouping logic from both consumers
