+++ apps/docsite/src/components/DocsShell.tsx
@@ -11,6 +11,7 @@ import type {ComponentEntry} from '../generated/componentRegistry';
 import type {PackageMeta} from '../generated/packageRegistry';
 import type {DocTopic} from '../generated/docsRegistry';
 import type {TemplateEntry} from '../generated/templateRegistry';
+import {groupComponents, type ComponentItem} from '../lib/groupComponents';
 
 interface DocsShellProps {
   children: React.ReactNode;
@@ -30,98 +31,14 @@ const foundationsSort = (a: DocTopic, b: DocTopic) => {
 
 // ── Component sidebar builder ──────────────────────────────────────────
 
-type SidebarItem =
-  | {type: 'entry'; name: string; href: string}
-  | {
-      type: 'group';
-      label: string;
-      entries: Array<{name: string; href: string}>;
-    };
+type SidebarItem = ComponentItem;
 
 function buildComponentSidebar(entries: ComponentEntry[]): {
   componentItems: SidebarItem[];
   utilities: Array<{name: string; href: string}>;
 } {
-  const utilities: Array<{name: string; href: string}> = [];
-  const groups = new Map<string, Array<{name: string; href: string}>>();
-  const ungrouped: Array<{name: string; href: string}> = [];
-
-  const parentDocsWithComponents = new Set<string>();
-  for (const e of entries) {
-    if (e.parentDoc && !e.name.startsWith('use') && !e.hidden) {
-      parentDocsWithComponents.add(e.parentDoc);
-    }
-  }
-
-  for (const entry of entries) {
-    if (entry.hidden) continue;
-    const isHook = entry.name.startsWith('use');
-
-    if (
-      entry.group === 'Utilities' ||
-      (isHook && !entry.parentDoc && entry.directory === 'hooks')
-    ) {
-      utilities.push({name: entry.name, href: `/components/${entry.name}`});
-      continue;
-    }
-    if (entry.group) {
-      if (!groups.has(entry.group)) groups.set(entry.group, []);
-      groups
-        .get(entry.group)!
-        .push({name: entry.name, href: `/components/${entry.name}`});
-      continue;
-    }
-    if (isHook && !entry.parentDoc && entry.directory !== 'hooks') {
-      const dir = entry.directory;
-      if (!groups.has(dir)) groups.set(dir, []);
-      groups
-        .get(dir)!
-        .push({name: entry.name, href: `/components/${entry.name}`});
-      continue;
-    }
-    if (
-      isHook &&
-      entry.parentDoc &&
-      parentDocsWithComponents.has(entry.parentDoc)
-    ) {
-      const parent = entry.parentDoc;
-      if (!groups.has(parent)) groups.set(parent, []);
-      groups
-        .get(parent)!
-        .push({name: entry.name, href: `/components/${entry.name}`});
-      continue;
-    }
-    if (isHook) {
-      utilities.push({name: entry.name, href: `/components/${entry.name}`});
-      continue;
-    }
-    ungrouped.push({name: entry.name, href: `/components/${entry.name}`});
-  }
-
-  const items: Array<{sortKey: string; item: SidebarItem}> = [];
-  for (const [label, members] of groups) {
-    members.sort((a, b) => a.name.localeCompare(b.name));
-    if (members.length === 1) {
-      items.push({
-        sortKey: members[0].name,
-        item: {type: 'entry', ...members[0]},
-      });
-    } else {
-      items.push({
-        sortKey: label,
-        item: {type: 'group', label, entries: members},
-      });
-    }
-  }
-  for (const entry of ungrouped) {
-    items.push({sortKey: entry.name, item: {type: 'entry', ...entry}});
-  }
-  items.sort((a, b) => a.sortKey.localeCompare(b.sortKey));
-
-  return {
-    componentItems: items.map(i => i.item),
-    utilities: utilities.sort((a, b) => a.name.localeCompare(b.name)),
-  };
+  const {items, utilities} = groupComponents(entries);
+  return {componentItems: items, utilities};
 }
 
 // ── Wordmark ───────────────────────────────────────────────────────────
